[core] Move executor.Iterable to graphql.Iterable.
Also add `NewMap{Keys,Values}Iterator`.
